+++ kipper/core/src/compiler/ast/nodes/declarations/type-declaration/class-declaration/class-declaration-semantics.ts
@@ -3,6 +3,9 @@
  * @since 0.11.0
  */
 import type { TypeDeclarationSemantics } from "../type-declaration-semantics";
+import type { ClassMemberDeclaration } from "./class-member-declaration";
+import { ClassMemberDeclarationSemantics } from "./class-member-declaration";
+import type { ClassConstructorDeclaration } from "./class-member-declaration/class-constructor-declaration/class-constructor-declaration";
 
 /**
  * Semantics for AST Node {@link ClassDeclaration}.
@@ -14,4 +17,16 @@ export interface ClassDeclarationSemantics extends TypeDeclarationSemantics {
 	 * @since 0.11.0
 	 */
 	identifier: string;
+
+	/**
+	 * The body of the class.
+	 * @since 0.11.0
+	 */
+	classMembers: Array<ClassMemberDeclaration>;
+
+	/**
+	 * The class constructor.
+	 * @since 0.11.0
+	 */
+	constructorDeclaration: ClassConstructorDeclaration | undefined;
 }